+++ plugins/main/public/components/health-check/container/health-check.container.tsx
@@ -129,6 +129,9 @@ function HealthCheckComponent() {
     window.location.href = getHttp().basePath.prepend(url);
   };
 
+  const thereAreErrors = Object.keys(checkErrors).length > 0;
+  const thereAreWarnings = Object.keys(checkWarnings).length > 0;
+
   useEffect(() => {
     if (appConfig.isReady && !checksInitiated.current) {
       checksInitiated.current = true;
@@ -140,6 +143,7 @@ function HealthCheckComponent() {
     // Redirect to app when all checks are ready
     Object.keys(checks).every(check => checksReady[check]) &&
       !isDebugMode &&
+      !thereAreWarnings &&
       (() =>
         setTimeout(
           redirectionPassHealthcheck,
@@ -196,8 +200,6 @@ function HealthCheckComponent() {
       ? getAssetURL(appConfig.data['customization.logo.healthcheck'])
       : getThemeAssetURL('logo.svg'),
   );
-  const thereAreErrors = Object.keys(checkErrors).length > 0;
-  const thereAreWarnings = Object.keys(checkWarnings).length > 0;
 
   const renderChecks = () => {
     const showLogButton = thereAreErrors || thereAreWarnings || isDebugMode;
@@ -350,7 +352,7 @@ function HealthCheckComponent() {
                 </RedirectAppLinks>
               </EuiFlexItem>
             )}
-            {isDebugMode &&
+            {(isDebugMode || thereAreWarnings) &&
               Object.keys(checks).every(check => checksReady[check]) && (
                 <EuiFlexItem grow={false}>
                   <EuiButton fill onClick={redirectionPassHealthcheck}>